Java-06 v2

+++ Task/src/main/java/dao/Dao.java
@@ -12,7 +12,7 @@ public interface Dao<T> {
 
     Account create(T account) throws IOException, SQLException, UnknownAccountException;
 
-    Account read(int id) throws IOException, UnknownAccountException, SQLException;
+    Account read(long id) throws IOException, UnknownAccountException, SQLException;
 
     Account update(T account) throws SQLException, IOException, NotEnoughMoneyException, UnknownAccountException;
